# improve mark handling performance for large selections