@ianstormtaylor 2017-07-12
Interesting, thanks @erquhart!